refactor(edge): reduce control api http module to route handling
